Harden `tag_id` overflow handling by enforcing platform-width parsing at API boundary (#36)
* fix: validate admin delete tag_id with platform uint width

Agent-Logs-Url: https://github.com/luoxiaojun1992/miniprogram/sessions/3f0646ed-3f93-4db5-84bd-327eb174cd4f

* test: make tag_id overflow test platform-aware

Agent-Logs-Url: https://github.com/luoxiaojun1992/miniprogram/sessions/3f0646ed-3f93-4db5-84bd-327eb174cd4f

* fix: add explicit uint bound check for admin delete user tag

Agent-Logs-Url: https://github.com/luoxiaojun1992/miniprogram/sessions/53e64066-936a-416d-a09e-c0153d7a7dbc


* test: make tag overflow test computed and 32-bit specific

Agent-Logs-Url: https://github.com/luoxiaojun1992/miniprogram/sessions/53e64066-936a-416d-a09e-c0153d7a7dbc


* refactor: use uint64 for delete user tag flow

Agent-Logs-Url: https://github.com/luoxiaojun1992/miniprogram/sessions/2912b27b-a3a2-42ef-984e-be78b773dcd8


---------

+++ internal/repository/interface.go
@@ -30,7 +30,7 @@ type AdminUserRepository interface {
 type UserTagRepository interface {
 	GetByUserID(ctx context.Context, userID uint64) ([]*entity.UserTag, error)
 	Create(ctx context.Context, tag *entity.UserTag) error
-	Delete(ctx context.Context, id uint) error
+	Delete(ctx context.Context, id uint64) error
 }
 
 // RoleRepository defines the interface for role data access.
